// src/navigation/CommonNavigator/types.ts
export type CommonNavigatorParamList = {
    EventScreen: any;
    EventFormScreen: any;

    ChatScreen: any;
    ClaimScreen: any;

    ProfileUserScreen: any;
    ProfileSettings: any;
    ProfileNotifications: any;
    ProfileReviews: any;
    ProfileApp: any;
    ProfileFormScreen: any;

    DebugScreen:any;
}